refactor(docker): split catalog and client boundaries
